requirements.txt
streamlit==1.31.1
openai==0.28.1
python-dotenv==1.0.0
protobuf==3.20 
